requirements.txt
PyQt5 >= 5.9.2
pydot >= 1.2.4
natsort >= 5.3.2
lark-parser >= 0.5.6
debug_tools >= 2.14
fastnumbers >= 2.0.0
